requirements.txt
fastapi
Jinja2
pydantic
PyYAML
pydantic-settings
python-dotenv
uvicorn[standard]
fastapi_msal
msal
msal-extensions
async-signals
websockets
itsdangerous
python-multipart
aiohttp
azure-core
azure-identity
azure-cosmos
azure-servicebus
azure-storage-blob
azure-monitor-opentelemetry
azure-monitor-events-extension
azure-messaging-webpubsubservice
pytest
